pyproject.toml
[tool.poetry]
name = "obviously"
version = "0.1.0"
description = ""
authors = ["Jay Chhatrala <jay.chhatrala@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
black = "^24.4.2"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
pytest = "^8.2.2"
pytest-django = "^4.8.0"
pytest-factoryboy = "^2.7.0"
freezegun = "^1.5.1"
djangorestframework = "^3.15.2"
django-filter = "^24.2"
django-vectordb = {extras = ["standard"], version = "^0.3.0"}
numpy = "<2"
sentencepiece = "^0.2.0"
hnswlib = "^0.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
